refactor: Rename bestPractices module to asyncComponentFlags, also check TS-Components
